+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added plot for viewing drug class counts and AMR genes.
 * Added `setup.py` install script.
 * Changed axis of timeline to percent instead of count.
+* Added *AMR Gene Family* and *Resistance Mechanism* as selection criteria and as figures.
 
 # 0.1.0